+++ src/test/java/LessonFour/TriangleTest.java
@@ -39,7 +39,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
     @DisplayName("Разносторонний треугольник - все стороны которого имеют разную длину.")
     public void countPerimeterVersatileTriangleTest() {
         Triangle triangle = new Triangle(6, 8, 10);
-        int perimeter = triangle.countArea();
+        double perimeter = triangle.countArea();
         assertEquals(24, perimeter);
     }
 
@@ -47,7 +47,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
     @DisplayName("Равносторонний треугольник - у которого все три стороны равны.")
     public void countPerimeterEquilateralTriangleTest() {
         Triangle triangle = new Triangle(6, 6, 6);
-        int perimeter = triangle.countArea();
+        double perimeter = triangle.countArea();
         assertEquals(15, perimeter);
     }
 
@@ -55,7 +55,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
     @DisplayName("Равнобедренный треугольник - у которого две стороны равны")
      public void countIsoscelesTriangleTest() {
         Triangle triangle = new Triangle(7, 7, 10);
-        int perimeter = triangle.countArea();
+       double perimeter = triangle.countArea();
         assertEquals(24, perimeter);
     }
 
@@ -69,7 +69,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
     @ParameterizedTest(name = "Периметр треугольника {0} равен {1}")
     @MethodSource("triangles")
     public void countPerimeterTriangleTest(Triangle triangle, int expectedPerimeter) {
-        int perimeter = triangle.countArea(); //Act
+        double perimeter = triangle.countArea(); //Act
         assertEquals(expectedPerimeter, perimeter); //Assert
     }
